server-node: Moved core code into 'core' directory
