+++ Gruntfile.js
@@ -83,7 +83,6 @@
           'src/*.js',
           'src/server/node/*.js',
           'src/server/node/**/*.js',
-          'src/server/node/node_modules/osjs/*.js',
           'src/client/javascript/*.js',
           'src/client/javascript/**/*.js',
           'src/packages/default/**/*.js',
@@ -165,7 +164,6 @@
           'src/*.js',
           'src/server/node/*.js',
           'src/server/node/**/*.js',
-          'src/server/node/node_modules/osjs/*.js',
           'src/client/javascript/*.js',
           'src/client/javascript/**/*.js',
           'src/packages/default/**/*.js',
